FREEDOM-186 Working alias clearing script

+++ custom/modules/aaslp_core/scripts/clr_alias.php
@@ -4,17 +4,25 @@
  * @file
  * Contains clr_alias.php.
  *
- * Clears node alias for slaveryandfreedomlaws.lib.unb.ca.
+ * Clears all node alias for legal articles in slaveryandfreedomlaws.lib.unb.ca.
  */
+
+$nids = \Drupal::entityQuery('node')
+  ->condition('type', 'legal_article')
+  ->execute();
 
 $path_alias_manager = \Drupal::entityTypeManager()->getStorage('path_alias');
 
-// Load all path alias for this node for es language.
-$alias_objects = $path_alias_manager->loadByProperties([
-  'path'     => '/node/' . $nid,
-  'langcode' => 'en',
-]);
+if (!empty($nids)) {
+  foreach ($nids as $nid) {
+    // Load all path alias for this node.
+    $alias_objects = $path_alias_manager->loadByProperties([
+      'path' => "/node/$nid",
+    ]);
 
-foreach ($alias_objects as $alias_object) {
-  $alias_object->delete();
+    foreach ($alias_objects as $alias_object) {
+      $alias_object->delete();
+      echo "[-] [node/{$nid}]->[Updated]\n";
+    }
+  }
 }
